+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## [2.0.0-alpha.0] - 2018-09-03
+## [Unreleased]
 ### Added
 - `Marshal` and `Unmarshal` functions.
 - Support for using custom structs embedding a `JWT` object as a JWT.
@@ -18,6 +18,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Removed
 - `Sign` and `Verify` functions.
 - JWT options object.
+- Base64 encoding and deconding functions.
 
 ## [1.1.0] - 2018-08-22
 ### Changed
@@ -109,7 +110,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - This package's source code, including examples and tests.
 - Go dep files.
 
-[2.0.0-alpha.0]: https://github.com/gbrlsnchs/jwt/compare/v1.1.0...v2.0.0-alpha.0
+[Unreleased]: https://github.com/gbrlsnchs/jwt/compare/v1.1.0...HEAD
 [1.1.0]: https://github.com/gbrlsnchs/jwt/compare/v1.0.2...v1.1.0
 [1.0.2]: https://github.com/gbrlsnchs/jwt/compare/v1.0.1...v1.0.2
 [1.0.1]: https://github.com/gbrlsnchs/jwt/compare/v1.0.0...v1.0.1